Move EnhancedComlumnsTemplate parts to separate components
